[rename] 커스텀 클래스 코드 이동 및 파일 삭제
