+++ BranchTalk/Custom/CustomLabel.swift
@@ -25,6 +25,24 @@ class CustomTitle1Label: UILabel {
     }
 }
 
+class CustomTitle2Label: UILabel {
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setLabel()
+    }
+    
+    private func setLabel() {
+        font = Font.title2Bold()
+        textColor = Colors.BrandBlack.CutsomColor
+        
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
+
 class CustomBodyLabel: UILabel {
     
     override init(frame: CGRect) {
